Add visibility animations, a new type of Object Animation
This animation simply shows and hides objects to produce the illusion of an animation. The implementation of this animation is very basic:
the "frame data" of the animation is other simple objects that are referenced by name, and after they are spawned during the animation,
their visibility its manipulated during the animation.