Start efforts to add dual paraboloid shadow mapping
as an alternative to cubemap for shadow map atlases first.
This is not working for some reason when testing for shadows in deferred_light fragment...
not sure if it's a problem in the deferred light fragment or before when rendering the shadow map faces, need to find some
implementation of this technique that can be inspected with renderdoc to help debugging if there is a problem with the matrices or the shaders
or something else.

Note:

- this technique requires manual projection done in the vertex shader, which all the other shadow map methods used by armory do not...
I couldn't find an easy way to "interchange" the shaders so the next simplest solution I could think of was just to add a new context: "shadowmapdp",
which from the name it can be deduced that it's a duplicate of "shadowmap", which seems to work fine and it wasn't too much disrupt since the coupling to
"shadowmap" is minimal, mostly in Inc and some classes in iron.

Reference for some bits of the implementation:
https://gamedev.stackexchange.com/questions/106615/dual-paraboloid-shadows-on-glsl
http://gamedevelop.eu/en/tutorials/dual-paraboloid-shadow-mapping.htm
http://graphicsrunner.blogspot.com/2008/07/dual-paraboloid-shadow-maps.html